@@ -23,14 +23,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.9",
-    "@angular/cdk": "13.3.7",
-    "@angular/common": "13.3.9",
-    "@angular/core": "13.3.9",
-    "@angular/forms": "13.3.9",
-    "@angular/platform-browser": "13.3.9",
-    "@angular/platform-browser-dynamic": "13.3.9",
-    "@angular/router": "13.3.9",
+    "@angular/animations": "14.0.1",
+    "@angular/cdk": "14.0.1",
+    "@angular/common": "14.0.1",
+    "@angular/core": "14.0.1",
+    "@angular/forms": "14.0.1",
+    "@angular/platform-browser": "14.0.1",
+    "@angular/platform-browser-dynamic": "14.0.1",
+    "@angular/router": "14.0.1",
     "@asymmetrik/ngx-leaflet": "13.0.2",
     "@capacitor-community/fcm": "2.0.2",
     "@capacitor-community/firebase-analytics": "1.0.1",
@@ -83,48 +83,48 @@
     "zone.js": "0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.3.6",
-    "@angular-eslint/builder": "13.2.1",
-    "@angular-eslint/eslint-plugin": "13.2.1",
-    "@angular-eslint/eslint-plugin-template": "13.2.1",
-    "@angular-eslint/schematics": "13.2.1",
-    "@angular-eslint/template-parser": "13.2.1",
-    "@angular/cli": "13.3.6",
-    "@angular/compiler": "13.3.9",
-    "@angular/compiler-cli": "13.3.9",
-    "@angular/language-service": "13.3.9",
+    "@angular-devkit/build-angular": "14.0.1",
+    "@angular-eslint/builder": "13.5.0",
+    "@angular-eslint/eslint-plugin": "13.5.0",
+    "@angular-eslint/eslint-plugin-template": "13.5.0",
+    "@angular-eslint/schematics": "13.5.0",
+    "@angular-eslint/template-parser": "13.5.0",
+    "@angular/cli": "14.0.1",
+    "@angular/compiler": "14.0.1",
+    "@angular/compiler-cli": "14.0.1",
+    "@angular/language-service": "14.0.1",
     "@capacitor/cli": "3.5.1",
-    "@cypress/schematic": "1.7.0",
+    "@cypress/schematic": "2.0.0",
     "@ionic/angular-toolkit": "6.1.0",
-    "@ngxs/devtools-plugin": "3.7.3",
-    "@ngxs/hmr-plugin": "3.7.3",
+    "@ngxs/devtools-plugin": "3.7.4",
+    "@ngxs/hmr-plugin": "3.7.4",
     "@types/jasmine": "4.0.3",
     "@types/jasminewd2": "2.0.10",
-    "@types/leaflet": "1.7.10",
-    "@types/node": "16.11.36",
-    "@typescript-eslint/eslint-plugin": "5.25.0",
-    "@typescript-eslint/parser": "5.25.0",
-    "@wdio/appium-service": "7.19.5",
-    "@wdio/cli": "7.19.7",
-    "@wdio/jasmine-framework": "7.19.7",
-    "@wdio/local-runner": "7.19.7",
-    "@wdio/spec-reporter": "7.19.7",
-    "chromedriver": "101.0.0",
+    "@types/leaflet": "1.7.11",
+    "@types/node": "16.11.39",
+    "@typescript-eslint/eslint-plugin": "5.27.1",
+    "@typescript-eslint/parser": "5.27.1",
+    "@wdio/appium-service": "7.20.1",
+    "@wdio/cli": "7.20.2",
+    "@wdio/jasmine-framework": "7.20.0",
+    "@wdio/local-runner": "7.20.2",
+    "@wdio/spec-reporter": "7.20.0",
+    "chromedriver": "102.0.0",
     "codelyzer": "6.0.2",
-    "cypress": "9.6.1",
-    "eslint": "8.16.0",
-    "jasmine-core": "4.1.1",
+    "cypress": "10.1.0",
+    "eslint": "8.17.0",
+    "jasmine-core": "4.2.0",
     "jasmine-spec-reporter": "7.0.0",
     "karma": "6.3.20",
     "karma-chrome-launcher": "3.1.1",
     "karma-coverage-istanbul-reporter": "3.0.3",
     "karma-jasmine": "5.0.1",
-    "karma-jasmine-html-reporter": "1.7.0",
+    "karma-jasmine-html-reporter": "2.0.0",
     "ng-openapi-gen": "0.21.2",
     "protractor": "7.0.0",
-    "ts-node": "10.7.0",
+    "ts-node": "10.8.1",
     "tslint": "6.1.3",
-    "typescript": "4.6.4",
+    "typescript": "4.7.3",
     "wdio-chromedriver-service": "7.3.2"
   },
   "description": "An Ionic project"
