@@ -1,18 +1,18 @@
 /* tslint:disable */
 /* eslint-disable */
-import { Injectable } from '@angular/core';
-import { HttpClient, HttpResponse } from '@angular/common/http';
-import { BaseService } from '../base-service';
-import { ApiConfiguration } from '../api-configuration';
-import { StrictHttpResponse } from '../strict-http-response';
-import { RequestBuilder } from '../request-builder';
-import { Observable } from 'rxjs';
-import { map, filter } from 'rxjs/operators';
+import {Injectable} from '@angular/core';
+import {HttpClient, HttpResponse} from '@angular/common/http';
+import {BaseService} from '../base-service';
+import {ApiConfiguration} from '../api-configuration';
+import {StrictHttpResponse} from '../strict-http-response';
+import {RequestBuilder} from '../request-builder';
+import {Observable} from 'rxjs';
+import {filter, map} from 'rxjs/operators';
 
-import { ClubEntry } from '../models/club-entry';
-import { MemberEntry } from '../models/member-entry';
-import { MemberResults } from '../models/member-results';
-import { TeamEntry } from '../models/team-entry';
+import {ClubEntry} from '../models/club-entry';
+import {MemberEntry} from '../models/member-entry';
+import {MemberResults} from '../models/member-results';
+import {TeamEntry} from '../models/team-entry';
 
 @Injectable({
   providedIn: 'root',
