@@ -1,17 +1,17 @@
 /* tslint:disable */
 /* eslint-disable */
-import { Injectable } from '@angular/core';
-import { HttpClient, HttpResponse } from '@angular/common/http';
-import { BaseService } from '../base-service';
-import { ApiConfiguration } from '../api-configuration';
-import { StrictHttpResponse } from '../strict-http-response';
-import { RequestBuilder } from '../request-builder';
-import { Observable } from 'rxjs';
-import { map, filter } from 'rxjs/operators';
+import {Injectable} from '@angular/core';
+import {HttpClient, HttpResponse} from '@angular/common/http';
+import {BaseService} from '../base-service';
+import {ApiConfiguration} from '../api-configuration';
+import {StrictHttpResponse} from '../strict-http-response';
+import {RequestBuilder} from '../request-builder';
+import {Observable} from 'rxjs';
+import {filter, map} from 'rxjs/operators';
 
-import { RegisterTournament } from '../models/register-tournament';
-import { TournamentEntry } from '../models/tournament-entry';
-import { TournamentSerieEntry } from '../models/tournament-serie-entry';
+import {RegisterTournament} from '../models/register-tournament';
+import {TournamentEntry} from '../models/tournament-entry';
+import {TournamentSerieEntry} from '../models/tournament-serie-entry';
 
 @Injectable({
   providedIn: 'root',
