@@ -1,8 +1,8 @@
 /* tslint:disable */
 /* eslint-disable */
-import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { ApiConfiguration } from './api-configuration';
+import {Injectable} from '@angular/core';
+import {HttpClient} from '@angular/common/http';
+import {ApiConfiguration} from './api-configuration';
 
 /**
  * Base class for services
