@@ -1,6 +1,7 @@
 /* tslint:disable */
 /* eslint-disable */
-import { TeamMatchesEntry } from './team-matches-entry';
+import {TeamMatchesEntry} from './team-matches-entry';
+
 export interface MatchEntryHistory {
   date: string;
   matchEntry: TeamMatchesEntry;
