@@ -1,7 +1,8 @@
 /* tslint:disable */
 /* eslint-disable */
-import { DoubleTeam } from './double-team';
-import { Player } from './player';
+import {DoubleTeam} from './double-team';
+import {Player} from './player';
+
 export interface Players {
   DoubleTeamCount: number;
   DoubleTeams?: Array<DoubleTeam>;
