@@ -1,7 +1,8 @@
 /* tslint:disable */
 /* eslint-disable */
-import { ClubEntry } from './club-entry';
-import { MemberEntry } from './member-entry';
+import {ClubEntry} from './club-entry';
+import {MemberEntry} from './member-entry';
+
 export interface RegistrationEntry {
   Club: ClubEntry;
   Member: MemberEntry;
