@@ -1,7 +1,8 @@
 /* tslint:disable */
 /* eslint-disable */
-import { MatchDetails } from './match-details';
-import { VenueEntry } from './venue-entry';
+import {MatchDetails} from './match-details';
+import {VenueEntry} from './venue-entry';
+
 export interface TeamMatchesEntry {
   AwayClub: string;
   AwayTeam: string;
