@@ -1,7 +1,8 @@
 /* tslint:disable */
 /* eslint-disable */
-import { RegistrationEntry } from './registration-entry';
-import { TournamentSerieResultEntry } from './tournament-serie-result-entry';
+import {RegistrationEntry} from './registration-entry';
+import {TournamentSerieResultEntry} from './tournament-serie-result-entry';
+
 export interface TournamentSerieEntry {
   Name: string;
   RegistrationCount?: number;
