@@ -1,6 +1,6 @@
-import { TestBed } from '@angular/core/testing';
+import {TestBed} from '@angular/core/testing';
 
-import { InAppPurchasesService } from './in-app-purchases.service';
+import {InAppPurchasesService} from './in-app-purchases.service';
 
 describe('InAppPurchasesService', () => {
   let service: InAppPurchasesService;
