@@ -1,6 +1,6 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import {ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { NotificationItemComponent } from './notification-item.component';
+import {NotificationItemComponent} from './notification-item.component';
 
 describe('NotificationItemComponent', () => {
   let component: NotificationItemComponent;
