@@ -1,6 +1,6 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import {ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { NotificationsComponent } from './notifications.component';
+import {NotificationsComponent} from './notifications.component';
 
 describe('NotificationsComponent', () => {
   let component: NotificationsComponent;
