@@ -1,6 +1,6 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import {ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { RotatioComponent } from './rotatio.component';
+import {RotatioComponent} from './rotatio.component';
 
 describe('RotatioComponent', () => {
   let component: RotatioComponent;
